requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Data & Image Processing
numpy>=1.24.0
Pillow>=9.5.0
gdown>=4.7.0

# Logging & Visualization
tensorboard>=2.13.0
tqdm>=4.65.0

# Web Demo
gradio>=4.0.0

# Evaluation
pytorch-fid>=0.3.0
scipy>=1.10.0

# Utilities
einops>=0.6.0

